move visualize code in entrypoint
